+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## v0.0.7 (20??-??-??)
 
+### Added
+
+* Support MLflow tracking (`border-mlflow-tracking`) (https://github.com/taku-y/border/issues/2).
+* Add candle agent (`border-candle-agent`)
+
 ### Changed
 
 * Take `self` in the signature of `push()` method of replay buffer (`border-core`)
-* Add candle agent (`border-candle-agent`)
 * Fix a bug in `MlpConfig` (`border-tch-agent`)
 
 ## v0.0.6 (2023-09-19)